refactor:MigrationProduct

+++ database/migrations/2021_02_12_112756_create_products_table.php
@@ -15,8 +15,8 @@ class CreateProductsTable extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
-            $table->integer('Quantity');
-            $table->integer('Cost');
+            $table->float('Quantity');
+            $table->float('Cost');
             $table->string('image');
         });
     }
